Support explicit JsType annotations field or class
We now support @JsType annotations at field or class level which
allow users to specify any type for that field or class. This is useful
when a type is serialized in a non-standard way and the user wants
to use a different TypeScript mapping than the one inferred by the
class structure.